Allow --clip argument when subcommand is omitted
The 'show' subcommand is the "default" subcommand (i.e. when no other
subcommand is given). clap doesn't appear to have a good way to specify
a "default" subcommand, so instead the 'clip' argument must be added
twice: both to the main app as well as to the 'show' subcommand. This is
redundant and I don't like it, and is a good candidate for a future
refactor.